NEXT-39968 - Rename Enumeration to Enum
