+++ tests/integration/Core/Framework/DataAbstractionLayer/Field/EnumField/TestIntegerEnum.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Shopware\Tests\Integration\Core\Framework\DataAbstractionLayer\Field\EnumField;
+
+/**
+ * @internal
+ */
+enum TestIntegerEnum: int
+{
+    case Zero = 0;
+    case One = 1;
+}
