+++ tests/integration/Core/Framework/DataAbstractionLayer/Field/EnumField/TestStringEnum.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Shopware\Tests\Integration\Core\Framework\DataAbstractionLayer\Field\EnumerationField;
+namespace Shopware\Tests\Integration\Core\Framework\DataAbstractionLayer\Field\EnumField;
 
 /**
  * @internal
  */
-enum TestStringEnumeration: string
+enum TestStringEnum: string
 {
     case Regular = 'string';
     case LeadingSpace = ' leading-space';
