--- tests/integration/Core/Framework/DataAbstractionLayer/Field/EnumerationField/TestIntegerEnumeration.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Shopware\Tests\Integration\Core\Framework\DataAbstractionLayer\Field\EnumerationField;
-
-/**
- * @internal
- */
-enum TestIntegerEnumeration: int
-{
-    case Zero = 0;
-    case One = 1;
-}
